Minor fixes

+++ generator.py
@@ -1,10 +1,8 @@
+import os
 import sys
 
-import os
-
-
 def to_kebab_case(snake_str):
-  return snake_str.replace("_", "-").lower()
+    return snake_str.replace("_", "-").lower()
 
 
 def to_camel_case(snake_str):
@@ -18,20 +16,24 @@ def to_pascal_case(snake_str):
 
 
 def code_table(app):
-  app_base = """
+    app_base = """
 import React, {{Component}} from 'react'
 
-import BasePage from 'base/components/BasePage'
+import BasePage from '../../base/components/BasePage'
 
 
-export default class {}Base extends Component {{
+export default class {0}Base extends Component {{
   render() {{
     return <BasePage children={{this.props.children}} />
   }}
 }}
-  """.format(to_pascal_case(app))
 
-  app_page = """
+{0}Base.propTypes = {{
+  children: React.PropTypes.array,
+}}
+    """.format(to_pascal_case(app))
+
+    app_page = """
 import React, {{Component}} from 'react'
 import {{StyleSheet, css}} from 'aphrodite'
 
@@ -53,12 +55,18 @@ export default class {1}Page extends Component {{
     )
   }}
 }}
-  """.format(to_camel_case(app), to_pascal_case(app))
 
-  app_model = """
+{1}Page.propTypes = {{
+  {0}Model: React.PropTypes.object,
+}}
+    """.format(to_camel_case(app), to_pascal_case(app))
+
+    app_model = """
 import {{observable, action}} from 'mobx'
+import remotedev from 'mobx-remotedev'
 
 
+@remotedev
 export default class {1}Model {{
   @observable info 
 
@@ -68,10 +76,10 @@ export default class {1}Model {{
 }}
 
 export const {0}Model = new {1}Model()
-  """.format(to_camel_case(app), to_pascal_case(app))
+    """.format(to_camel_case(app), to_pascal_case(app))
 
 
-  router = """
+    router = """
 import React from 'react'
 import {{ Route, IndexRoute }} from 'react-router'
 
@@ -85,57 +93,54 @@ const {1}Routes = (
 )
 
 export default {1}Routes
-  """.format(to_pascal_case(app), to_camel_case(app), to_kebab_case(app).lower())
+    """.format(to_pascal_case(app), to_camel_case(app), to_kebab_case(app).lower())
 
-  return {
-    "routes.js": router,
-    "models/{}.js".format(to_kebab_case(app)): app_model,
-    "components/{}Base.jsx".format(to_pascal_case(app)): app_base,
-    "components/{}Page.jsx".format(to_pascal_case(app)): app_page,
-  }
+    return {
+        "routes.js": router,
+        "models/{}.js".format(to_kebab_case(app)): app_model,
+        "components/{}Base.jsx".format(to_pascal_case(app)): app_base,
+        "components/{}Page.jsx".format(to_pascal_case(app)): app_page,
+    }
 
 
 def main():
-  try:
-    arg = sys.argv[1]
-  except:
-    print("Please Enter an app name")
-    return
-
-  location_prefix = "src"
-  app_folder = to_kebab_case(arg.lower())
-
-  folders = ["models", "components"]
-
-  files = ["routes.js", "models/{}.js".format(to_kebab_case(arg)), 
-           "components/{}Base.jsx".format(to_pascal_case(arg)),
-           "components/{}Page.jsx".format(to_pascal_case(arg))]
-  
-  try:
-    os.system("mkdir src/{}".format(app_folder))
-  except:
-    print("App already exists, can't override'")
-    return
-
-  for folder in folders:
     try:
-      os.system("mkdir src/{}/{}".format(app_folder, folder))
-      print(arg)
-      code_files = code_table(arg)
-    except AssertionError:
-      print("An error occured")
-      os.system("rm -Rf src/{}".format(app_folder))
-      return
+        arg = sys.argv[1]
+    except IndexError as _:
+        print("Please Enter an app name")
+        return
 
-  for f in files:
-    with file("src/{}/{}".format(app_folder, f), "w") as fin:
-      fin.write(code_files[f])
+    app_folder = to_kebab_case(arg.lower())
 
-  print("\n=== === Please add the following lines in the appropriate location in src/App.js === ===\n")
-  print("\n/* ------ {} APP --- --- */\n".format(arg.upper()))
-  print("import {}Routes from '{}/routes'".format(to_camel_case(arg), to_kebab_case(arg)))
-  print("import {{{0}Model}} from '{1}/models/{1}'".format(to_camel_case(arg), to_kebab_case(arg)))
-  print("\n=== === Add the stores and routers in appropriate locations as marked === ===\n")
+    folders = ["models", "components"]
+
+    files = ["routes.js", "models/{}.js".format(to_kebab_case(arg)), 
+             "components/{}Base.jsx".format(to_pascal_case(arg)),
+             "components/{}Page.jsx".format(to_pascal_case(arg))]
+  
+    res = os.system("mkdir src/{}".format(app_folder))
+    if res == 256:
+        print("App already exists, can't override")
+        return
+
+    for folder in folders:
+        res = os.system("mkdir src/{}/{}".format(app_folder, folder))
+
+        if res == 256:
+            print("An error occured")
+            os.system("rm -Rf src/{}".format(app_folder))  
+        else:
+            code_files = code_table(arg)
+
+    for f in files:
+        with file("src/{}/{}".format(app_folder, f), "w") as fin:
+            fin.write(code_files[f])
+
+    print("\n=== === Please add the following lines in the appropriate location in src/App.js === ===\n")
+    print("\n/* ------ {} APP --- --- */\n".format(arg.upper()))
+    print("import {}Routes from './{}/routes'".format(to_camel_case(arg), to_kebab_case(arg)))
+    print("import {{{0}Model}} from './{1}/models/{1}'".format(to_camel_case(arg), to_kebab_case(arg)))
+    print("\n=== === Add the stores and routers in appropriate locations as marked === ===\n")
 
 if __name__ == "__main__":
-  main()
+    main()
